# Increase sector root performance