--- a/internal/blake2b/gen.go
+++ b/internal/blake2b/gen.go
@@ -230,6 +230,8 @@ func genHashBlocksAVX2() {
 		}
 	}
 
+	Comment("Clear the upper YMM registers to avoid performance penalties")
+	VZEROUPPER()
 	RET()
 }